# openapi/api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mock Cloud API",
    "description": "\n    ## Mock Cloud API\n    \n    A comprehensive mock cloud backend API for testing, development, and SDK generation.\n    \n    ### Features\n    * **Virtual Machines** - Create, manage, and monitor VMs\n    * **Volumes** - Block storage management with VM attachment\n    * **Security Groups** - Network security and access control\n    * **Environments** - Network isolation with CIDR management\n    * **Async Operations** - Background task processing with Celery\n    \n    ### Use Cases\n    * **Development & Testing** - Mock cloud infrastructure\n    * **SDK Development** - Generate client libraries\n    * **Terraform Providers** - Infrastructure as code testing\n    * **CI/CD Pipelines** - Automated testing environments\n    \n    ### Authentication\n    No authentication required - suitable for development and testing.\n    \n    ### Rate Limiting\n    No rate limiting applied - suitable for load testing.\n    \n    ### SDK Generation\n    This API is designed to work seamlessly with OpenAPI code generators:\n    * Python SDK generation\n    * Go SDK generation  \n    * Node.js SDK generation\n    * Terraform provider development\n    ",
    "contact": {
      "name": "Mock Cloud API Support",
      "email": "support@mockcloud.local"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
...
  "paths": {
    "/api/v1/environments/": {
      "post": {
        "tags": [
          "environments",
          "environments"
        ],
        "summary": "Create Environment",
        "description": "Create a new network environment.\n\nCreates an isolated network environment with a specified CIDR block.\nAll resources (VMs, volumes) created within this environment will be\nnetwork-isolated from other environments.\n\n**Use Cases:**\n* Development environment isolation\n* Multi-tenant infrastructure\n* Network security testing\n\n**Example:**\n```json\n{\n  \"name\": \"prod-web\",\n  \"network_cidr\": \"10.100.0.0/16\",\n  \"description\": \"Production web environment\"\n}\n```",
        "operationId": "create_environment_api_v1_environments__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "environments",
          "environments"
        ],
        "summary": "Get All Environments",
        "description": "Get all environments with pagination.\n\nRetrieves a paginated list of all network environments.\nUseful for listing available environments and implementing\npagination in client applications.\n\n**Parameters:**\n* `skip`: Number of environments to skip (for pagination)\n* `limit`: Maximum number of environments to return\n\n**Returns:**\nPaginated list of environments with total count",
        "operationId": "get_all_environments_api_v1_environments__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/environments/{environment_id}": {
      "get": {
        "tags": [
          "environments",
          "environments"
        ],
        "summary": "Get Environment",
        "description": "Get a specific environment by ID.\n\nRetrieves detailed information about a network environment including\nits CIDR block, creation time, and associated resources.\n\n**Parameters:**\n* `environment_id`: Unique identifier for the environment\n\n**Returns:**\nComplete environment information with metadata",
        "operationId": "get_environment_api_v1_environments__environment_id__get",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Environment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "environments",
          "environments"
        ],
        "summary": "Delete Environment",
        "description": "Delete an environment by ID.\n\nRemoves a network environment and all its associated resources.\nThis operation is irreversible and will affect all VMs and volumes\nwithin the environment.\n\n**Parameters:**\n* `environment_id`: Unique identifier for the environment to delete\n\n**Returns:**\nSuccess message confirming deletion\n\n**Note:**\nThis operation will fail if there are active resources in the environment.",
        "operationId": "delete_environment_api_v1_environments__environment_id__delete",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Environment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EnvironmentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "network_cidr": {
            "type": "string",
            "title": "Network Cidr",
            "default": "10.0.0.0/16"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "EnvironmentCreate"
      },
      "EnvironmentListResponse": {
        "properties": {
          "environments": {
            "items": {
              "$ref": "#/components/schemas/EnvironmentResponse"
            },
            "type": "array",
            "title": "Environments"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "environments",
          "total"
        ],
        "title": "EnvironmentListResponse"
      },
      "EnvironmentResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "network_cidr": {
            "type": "string",
            "title": "Network Cidr",
            "default": "10.0.0.0/16"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at"
        ],
        "title": "EnvironmentResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}